deleted save command
