+++ apps/mocksi-lite/content/EditMode/utils.ts
@@ -6,3 +6,33 @@ export const applyStyles = (
 		element.style.setProperty(key, value);
 	}
 };
+
+export const buildQuerySelector = (
+	parentElement: HTMLElement,
+	newValue: string,
+) => {
+	const { localName, id, classList } = parentElement;
+	let keyToSave = localName;
+	if (id) {
+		keyToSave += `#${id}`;
+	}
+	if (classList.length) {
+		const filteredClasses = [...classList].filter((cls) => !cls.includes(":"));
+		keyToSave += `.${filteredClasses.join(".")}`;
+	}
+	let elements: NodeListOf<Element>;
+	try {
+		elements = document.querySelectorAll(keyToSave);
+	} catch (e: unknown) {
+		if (e instanceof Error) {
+			console.error(`Error querying selector ${keyToSave}: ${e}`);
+		}
+		return keyToSave;
+	}
+
+	if (elements.length) {
+		keyToSave += `[${[...elements].indexOf(parentElement)}]`;
+	}
+	keyToSave += `{${newValue}}`;
+	return keyToSave;
+};
